moved function to `convert.simplex` + created file for test
