+++ tests/core/test_simplicialcomplex.py
@@ -595,26 +595,3 @@ def test_issue_445(edgelist1):
     assert 1 not in S
     assert 0 not in S.edges
     assert S._edge == xgi.dual_dict(S._node)
-
-
-def test_k_skeleton(edgelist1, edgelist2):
-    S = xgi.SimplicialComplex(edgelist1)
-    S1 = S.k_skeleton(1, in_place=False)
-    edges_skeleton = [
-        frozenset({4}),
-        frozenset({5, 6}),
-        frozenset({1, 2}),
-        frozenset({8, 7}),
-        frozenset({2, 3}),
-        frozenset({6, 7}),
-        frozenset({8, 6}),
-        frozenset({1, 3}),
-    ]
-    assert set(S1.edges.members()) == set(edges_skeleton)
-
-    with pytest.raises(XGIError):
-        S.k_skeleton(5, in_place=True)
-
-    S2 = xgi.SimplicialComplex(edgelist2)
-    S2.k_skeleton(0, in_place=True)
-    assert S2.num_edges == 0
